Update and rename Ejercicio3.ps1 to Batch User Creation Script in PowerShell.ps1
This PowerShell script automates the creation of multiple local user accounts on a Windows system. It reads a list of usernames from a text file (users.txt) located on the user's desktop. The script prompts the administrator to input a secure password, which is applied to all the users being created. For each username in the file, the script creates a new local user account with the specified password. This is a useful tool for system administrators who need to set up multiple accounts quickly and consistently.