Update and rename ejerc3.ps1 to "Count .txt Files in Directory Script.ps1
This PowerShell script counts the number of .txt files in the current directory. It iterates through all files with the .txt extension using the Get-ChildItem cmdlet and maintains a count of these files. Finally, it outputs the total count to the console. This script is a quick and efficient way to determine the number of text files in a specified folder.